Conflicts:
	bin/cloud-local.sh

+++ bin/cloud-local.sh
@@ -46,11 +46,18 @@ function download_packages() {
 
   local maven=${pkg_src_maven}
 
-  declare -a urls=("${maven}/org/apache/accumulo/accumulo/${pkg_accumulo_ver}/accumulo-${pkg_accumulo_ver}-bin.tar.gz"
-                   "${mirror}/hadoop/common/hadoop-${pkg_hadoop_ver}/hadoop-${pkg_hadoop_ver}.tar.gz"
+  declare -a urls=("${mirror}/hadoop/common/hadoop-${pkg_hadoop_ver}/hadoop-${pkg_hadoop_ver}.tar.gz"
                    "${mirror}/zookeeper/zookeeper-${pkg_zookeeper_ver}/zookeeper-${pkg_zookeeper_ver}.tar.gz"
                    "${mirror}/kafka/${pkg_kafka_ver}/kafka_${pkg_kafka_scala_ver}-${pkg_kafka_ver}.tgz"
                    "${mirror}/spark/spark-${pkg_spark_ver}/spark-${pkg_spark_ver}-bin-without-hadoop.tgz")
+
+  if [[ "$acc_enable" -eq 1 ]]; then
+    urls=("${urls[@]}", "${maven}/org/apache/accumulo/accumulo/${pkg_accumulo_ver}/accumulo-${pkg_accumulo_ver}-bin.tar.gz")
+  fi
+
+  if [[ "$hbase_enable" -eq 1 ]]; then
+    urls=("${urls[@]}", "${mirror}/hbase/${pkg_hbase_ver}/hbase-${pkg_hbase_ver}-bin.tar.gz")
+  fi
 
   for x in "${urls[@]}"; do
       fname=$(basename "$x");
@@ -69,7 +76,8 @@ function unpackage {
 
   echo "Unpackaging software..."
   (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/zookeeper-${pkg_zookeeper_ver}.tar.gz") && echo "Unpacked zookeeper"
-  (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/accumulo-${pkg_accumulo_ver}-bin.tar.gz") && echo "Unpacked accumulo"
+  [[ "$acc_enable" -eq 1 ]] && (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/accumulo-${pkg_accumulo_ver}-bin.tar.gz") && echo "Unpacked accumulo"
+  [[ "$hbase_enable" -eq 1 ]] && (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/hbase-${pkg_hbase_ver}-bin.tar.gz") && echo "Unpacked hbase"
   (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/hadoop-${pkg_hadoop_ver}.tar.gz") && echo "Unpacked hadoop"
   (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/kafka_${pkg_kafka_scala_ver}-${pkg_kafka_ver}.tgz") && echo "Unpacked kafka"
   (cd -P "${CLOUD_HOME}" && tar $targs "${CLOUD_HOME}/pkg/spark-${pkg_spark_ver}-bin-without-hadoop.tgz") && echo "Unpacked spark"
@@ -80,19 +88,26 @@ function configure {
   cp -r ${CLOUD_HOME}/templates/* ${CLOUD_HOME}/tmp/staging/
 
   # accumulo config before substitutions
-  cp $ACCUMULO_HOME/conf/examples/3GB/standalone/* $ACCUMULO_HOME/conf/
+  [[ "$acc_enable" -eq 1 ]] && cp $ACCUMULO_HOME/conf/examples/3GB/standalone/* $ACCUMULO_HOME/conf/
 
   ## Substitute env vars
   sed -i~orig "s#LOCAL_CLOUD_PREFIX#${CLOUD_HOME}#;s#CLOUD_LOCAL_HOSTNAME#${CL_HOSTNAME}#;s#CLOUD_LOCAL_BIND_ADDRESS#${CL_BIND_ADDRESS}#" ${CLOUD_HOME}/tmp/staging/*/*
- 
-  # accumulo config
-  # make accumulo bind to all network interfaces (so you can see the monitor from other boxes)
-  sed -i~orig "s/\# export ACCUMULO_MONITOR_BIND_ALL=\"true\"/export ACCUMULO_MONITOR_BIND_ALL=\"true\"/" "${ACCUMULO_HOME}/conf/accumulo-env.sh"
-  echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/gc
-  echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/masters
-  echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/slaves
-  echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/monitor
-  echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/tracers
+
+  if [[ "$acc_enable" -eq 1 ]]; then
+    # accumulo config
+    # make accumulo bind to all network interfaces (so you can see the monitor from other boxes)
+    sed -i~orig "s/\# export ACCUMULO_MONITOR_BIND_ALL=\"true\"/export ACCUMULO_MONITOR_BIND_ALL=\"true\"/" "${ACCUMULO_HOME}/conf/accumulo-env.sh"
+    echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/gc
+    echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/masters
+    echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/slaves
+    echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/monitor
+    echo "${CL_HOSTNAME}" > ${ACCUMULO_HOME}/conf/tracers
+  fi
+
+  if [[ "$hbase_enable" -eq 1 ]]; then
+    sed -i~orig "s/\# export HBASE_MANAGES_ZK=true/export HBASE_MANAGES_ZK=false/" "${HBASE_HOME}/conf/hbase-env.sh"
+    echo "${CL_HOSTNAME}" > ${HBASE_HOME}/conf/regionservers
+  fi
 
   # hadoop slaves file
   echo "${CL_HOSTNAME}" > ${CLOUD_HOME}/tmp/staging/hadoop/slaves
@@ -105,8 +120,9 @@ function configure {
   test -d $KAFKA_HOME/config || mkdir $KAFKA_HOME/config
   cp ${CLOUD_HOME}/tmp/staging/hadoop/* $HADOOP_CONF_DIR/
   cp ${CLOUD_HOME}/tmp/staging/zookeeper/* $ZOOKEEPER_HOME/conf/
-  cp ${CLOUD_HOME}/tmp/staging/accumulo/* $ACCUMULO_HOME/conf/
   cp ${CLOUD_HOME}/tmp/staging/kafka/* $KAFKA_HOME/config/
+  [[ "$acc_enable" -eq 1 ]] && cp ${CLOUD_HOME}/tmp/staging/accumulo/* ${ACCUMULO_HOME}/conf/
+  [[ "$hbase_enable" -eq 1 ]] && cp ${CLOUD_HOME}/tmp/staging/hbase/* ${HBASE_HOME}/conf/
 
   # If Spark doesn't have log4j settings, use the Spark defaults
   test -f $SPARK_HOME/conf/log4j.properties || cp $SPARK_HOME/conf/log4j.properties.template $SPARK_HOME/conf/log4j.properties
@@ -153,17 +169,25 @@ function start_first_time {
   
   # sleep 
   sleep 5
-
-  # init accumulo
-  echo "Initializing accumulo"
-  $ACCUMULO_HOME/bin/accumulo init --instance-name $cl_acc_inst_name --password $cl_acc_inst_pass
   
-  # sleep
-  sleep 5
+  if [[ "$acc_enable" -eq 1 ]]; then
+    # init accumulo
+    echo "Initializing accumulo"
+    $ACCUMULO_HOME/bin/accumulo init --instance-name $cl_acc_inst_name --password $cl_acc_inst_pass
 
-  # starting accumulo
-  echo "starting accumulo..."
-  $ACCUMULO_HOME/bin/start-all.sh
+    # sleep
+    sleep 5
+
+    # starting accumulo
+    echo "starting accumulo..."
+    $ACCUMULO_HOME/bin/start-all.sh
+  fi
+
+  if [[ "$hbase_enable" -eq 1 ]]; then
+    # start hbase
+    echo "starting hbase..."
+    ${HBASE_HOME}/bin/start-hbase.sh
+  fi
 }
 
 function start_cloud {
@@ -173,6 +197,7 @@ function start_cloud {
   start_kafka
   start_hadoop
   start_accumulo
+  start_hbase
 }
 
 function start_zk {
@@ -183,8 +208,10 @@ function start_zk {
 
 function start_kafka {
   # start kafka
-  echo "Starting kafka..."
-  $KAFKA_HOME/bin/kafka-server-start.sh -daemon $KAFKA_HOME/config/server.properties
+  if [[ "$kafka_enable" -eq 1 ]]; then
+	echo "Starting kafka..."
+    $KAFKA_HOME/bin/kafka-server-start.sh -daemon $KAFKA_HOME/config/server.properties
+  fi
 }
 
 function start_hadoop {
@@ -201,9 +228,19 @@ function start_hadoop {
 }
 
 function start_accumulo {
-  # start accumulo
-  echo "starting accumulo..."
-  $ACCUMULO_HOME/bin/start-all.sh
+  if [[ "$acc_enable" -eq 1 ]]; then
+    # starting accumulo
+    echo "starting accumulo..."
+    $ACCUMULO_HOME/bin/start-all.sh
+  fi
+}
+
+function start_hbase {
+  if [[ "$hbase_enable" -eq 1 ]]; then
+    # start hbase
+    echo "starting hbase..."
+    ${HBASE_HOME}/bin/start-hbase.sh
+  fi
 }
 
 function start_yarn {
@@ -214,18 +251,30 @@ function start_yarn {
 function stop_cloud {
   stop_kafka
   stop_accumulo
+  stop_hbase
   stop_hadoop
   stop_zk
 }
 
 function stop_kafka {
-  echo "Stopping kafka..."
-  $KAFKA_HOME/bin/kafka-server-stop.sh
+  if [[ "$kafka_enable" -eq 1 ]]; then
+	echo "Stopping kafka..."
+	$KAFKA_HOME/bin/kafka-server-stop.sh
+ fi
 }
 
 function stop_accumulo {
-  echo "Stopping accumulo..."
-  $ACCUMULO_HOME/bin/stop-all.sh
+  if [[ "$acc_enable" -eq 1 ]]; then
+    echo "Stopping accumulo..."
+    $ACCUMULO_HOME/bin/stop-all.sh
+  fi
+}
+
+function stop_hbase {
+  if [[ "$hbase_enable" -eq 1 ]]; then
+    echo "Stopping hbase..."
+    ${HBASE_HOME}/bin/stop-hbase.sh
+  fi
 }
 
 function stop_hadoop {
@@ -247,7 +296,8 @@ function stop_yarn {
 }
 
 function clear_sw {
-  rm -rf "${CLOUD_HOME}/accumulo-${pkg_accumulo_ver}"
+  [[ "$acc_enable" -eq 1 ]] && rm -rf "${CLOUD_HOME}/accumulo-${pkg_accumulo_ver}"
+  [[ "$hbase_enable" -eq 1 ]] && rm -rf "${CLOUD_HOME}/hbase-${pkg_hbase_ver}"
   rm -rf "${CLOUD_HOME}/hadoop-${pkg_hadoop_ver}"
   rm -rf "${CLOUD_HOME}/zookeeper-${pkg_zookeeper_ver}"
   rm -rf "${CLOUD_HOME}/kafka_${pkg_kafka_scala_ver}-${pkg_kafka_ver}"
@@ -268,7 +318,12 @@ function clear_data {
 }
 
 function show_help {
-  echo "Provide 1 command: (init|start|stop|reconfigure|reyarn|clean|help)"
+  echo "Single argument commands: (init|start|stop|reconfigure|reyarn|clean|help)"
+  echo "Other commands to stop/start individual services:"
+  echo "  hadoop (start|stop)"
+  echo "  accumulo (start|stop)"
+  echo "  hbase (start|stop)"
+  echo "  kafka (start|stop)"
 }
 
 if [ "$#" -eq 0 ]; then
@@ -308,6 +363,12 @@ elif [[ $1 == 'kafka' ]]; then
   elif [[ $2 == 'stop' ]]; then
 	stop_kafka
   fi
+elif [[ $1 == 'hbase' ]]; then
+  if [[ $2 == 'start' ]]; then
+    start_hbase
+  elif [[ $2 == 'stop' ]]; then
+    stop_hbase
+  fi	
 elif [[ $1 == 'stop' ]]; then
   echo "Stopping Cloud..."
   stop_cloud
